+++ src/package.json
@@ -21,12 +21,10 @@
         "postcss": "^8.1.14",
         "react": "^18.2.0",
         "react-dom": "^18.1.0",
-        "react-refresh": "^0.13.0",
-        "ts-loader": "^9.3.0",
-        "typescript": "^4.7.2"
+        "react-refresh": "^0.13.0"
     },
     "dependencies": {
-        "@react-three/drei": "^9.13.2",
+        "@react-three/drei": "^9.14.3",
         "@react-three/fiber": "^8.0.27",
         "@types/three": "^0.141.0",
         "browser-sync-webpack-plugin": "^2.3.0",
