+++ src/resources/js/app.js
@@ -1,0 +1,22 @@
+import "../css/app.css";
+import { QueryClient, QueryClientProvider } from "react-query";
+import ReactDOM from "react-dom";
+import { ReactQueryDevtools } from "react-query/devtools";
+import CanvasContainer from "./component/CanvasContainer";
+
+const queryClient = new QueryClient();
+
+export function App() {
+    return (
+        <QueryClientProvider client={queryClient}>
+            <CanvasContainer />
+            {process.env.NODE_ENV === "development" && (
+                <ReactQueryDevtools initialIsOpen={false} />
+            )}
+        </QueryClientProvider>
+    );
+}
+
+if (document.getElementById("app")) {
+    ReactDOM.render(<App />, document.getElementById("app"));
+}
