--- src/resources/ts/container/EnhancedFrame.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-import { Canvas, useFrame } from '@react-three/fiber'
-import Frame from '../component/Frame'
-const EnhancedFrame: React.FC = () => {
-    return (
-        <>
-            <Frame />
-        </>
-    )
-}
-export default EnhancedFrame
